Not for building - Working through AvgSessionLength

+++ SuiteCareers/SuiteCareers/ViewModels/DashboardVM.cs
@@ -11,14 +11,14 @@ namespace SuiteCareers.ViewModels
         /*Snapshot Bar data*/
         public int NewUser { get; set; }    /*done*/
         public int TotalUsers { get; set; } /*done*/
-        public int AvgSessionLength { get; set; }
+        public double AvgSessionLength { get; set; }
         public int TotalSessions { get; set; }  /*done*/
         public int QuestionsAnswered { get; set; }
-        public int NewQuestions { get; set; }
+        public int NewQuestions { get; set; } /*done*/
 
         /*Active info data*/
         public int ActiveUsers { get; set; }
-        public int ActiveSessions { get; set; }
+        public int ActiveSessions { get; set; } /*done*/
 
         public IEnumerable<Session>? RecentSessions { get; set; }
         public IEnumerable<Question>? TopQuestions { get; set; }
